Refactor ProjectAndFolderCards to always render seperate panels for a11y
