+++ front/app/components/ProjectAndFolderCards/components/ProjectsList.tsx
@@ -1,9 +1,7 @@
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
 
 // components
-import ProjectCard from 'components/ProjectCard';
-import Outlet from 'components/Outlet';
+import ProjectsTabPanel from './ProjectsTabPanel';
 
 // hooks
 import { useWindowSize } from '@citizenlab/cl2-component-library';
@@ -16,30 +14,27 @@ import { PublicationTab } from '../';
 // utils
 import getCardSizes from './getCardSizes';
 import { isEqual } from 'lodash-es';
-import { getTabId, getTabPanelId } from './Topbar/Tabs';
 
 export type TCardSize = 'small' | 'medium' | 'large';
 
-const Container = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-`;
-
-const MockProjectCard = styled.div`
-  height: 1px;
-  background: transparent;
-  width: calc(33% - 12px);
-`;
-
-interface Props {
+export interface BaseProps {
   currentTab: PublicationTab;
   list: IAdminPublicationContent[];
   layout: TLayout;
   hasMore: boolean;
 }
 
-const ProjectsList = ({ currentTab, list, layout, hasMore }: Props) => {
+interface Props extends BaseProps {
+  availableTabs: PublicationTab[];
+}
+
+const ProjectsList = ({
+  currentTab,
+  availableTabs,
+  list,
+  layout,
+  hasMore,
+}: Props) => {
   const { windowWidth } = useWindowSize();
 
   const [cardSizes, setCardSizes] = useState<TCardSize[]>([]);
@@ -56,62 +51,18 @@ const ProjectsList = ({ currentTab, list, layout, hasMore }: Props) => {
   }, [list.length, layout]);
 
   return (
-    <Container id="e2e-projects-list">
-      <div
-        id={getTabPanelId(currentTab)}
-        role="tabpanel"
-        tabIndex={0}
-        aria-labelledby={getTabId(currentTab)}
-      >
-        {list.map((item: IAdminPublicationContent, index: number) => {
-          const projectOrFolderId = item.publicationId;
-          const projectOrFolderType = item.publicationType;
-          const size =
-            layout === 'dynamic'
-              ? cardSizes[index]
-              : layout === 'threecolumns'
-              ? 'small'
-              : 'medium';
-
-          return (
-            <React.Fragment key={index}>
-              {projectOrFolderType === 'project' && (
-                <ProjectCard
-                  projectId={projectOrFolderId}
-                  size={size}
-                  layout={layout}
-                />
-              )}
-              <Outlet
-                id="app.components.ProjectAndFolderCards.card"
-                publication={item}
-                size={size}
-                layout={layout}
-              />
-            </React.Fragment>
-          );
-        })}
-
-        {/*
-          A bit of a hack (but the most elegant one I could think of) to
-          make the 3-column layout work for the last row of project cards when
-          the total amount of projects is not divisible by 3 and therefore doesn't take up the full row width.
-          Ideally would have been solved with CSS grid, but... IE11
-        */}
-        {!hasMore &&
-          (layout === 'threecolumns' || list.length > 6) &&
-          (list.length + 1) % 3 === 0 && <MockProjectCard className={layout} />}
-
-        {!hasMore &&
-          (layout === 'threecolumns' || list.length > 6) &&
-          (list.length - 1) % 3 === 0 && (
-            <>
-              <MockProjectCard className={layout} />
-              <MockProjectCard className={layout} />
-            </>
-          )}
-      </div>
-    </Container>
+    <div id="e2e-projects-list">
+      {availableTabs.map((tab) => (
+        <ProjectsTabPanel
+          currentTab={currentTab}
+          tab={tab}
+          list={list}
+          layout={layout}
+          cardSizes={cardSizes}
+          hasMore={hasMore}
+        />
+      ))}
+    </div>
   );
 };
